WIP: Re-org
While looking for examples of tweaks to the tokyonight colorscheme, I
stumbled across [Ricardo Rodríguez](https://github.com/magidc/nvim-config)
configs. The way he's got the colorscheme configuration tweaks bundled up
nicely into a theme module is pretty much what I was wanting.

While looking over the rest of the repo, I really liked how it was
organized with plugins defined in the init of a `plugins` directory and any
configuration for plugins in a `plugins/configs`. LSP is still a bit of an
oddball, c'est la vie.

* [x] Reconciled list of plugins and configuration with my own, there is
      still one plugin I need to review, add and configure and I left a
      FIXME to do that: SmiteshP/nvim-navic (successor to SmiteshP/nvim-gps)
* [x] Reconcile themes
* [x] Reconcile autocmds
* [x] Reconcile commands
* [x] Reconcile mappings
      ish.. I'll adopt these mapping on trial, including using <space>
      as leader, I usually use comma but have been wanting to try
      switching to space so we'll see how it goes.
* [x] Reconcile settings/options
* [ ] Reconcile LSP/DAP/null-ls